// EchoLog/src/generated/models/Cr4c3_auditlogsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Cr4c3_auditlogsBase {
  cr4c3_action?: number;
  cr4c3_actorrole?: string;
  cr4c3_auditlogid?: string;
  createdon?: string;
  cr4c3_description?: string;
  cr4c3_entityid?: string;
  cr4c3_entitytype?: string;
  cr4c3_fieldchanged?: string;
  importsequencenumber?: number;
  modifiedon?: string;
  cr4c3_newvalue?: string;
  cr4c3_oldvalue?: string;
  overriddencreatedon?: string;
  statuscode?: number;
  statecode?: number;
  timezoneruleversionnumber?: number;
  cr4c3_timestamp?: string;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
  _cr4c3_action_label?: _cr4c3_action_label_cr4c3_action_label;
  _cr4c3_actor_value?: string;
  _cr4c3_actor_type?: _cr4c3_actor_type_cr4c3_actor_type;
  _createdby_value?: string;
  _createdonbehalfby_value?: string;
  _createdonbehalfby_type?: _createdonbehalfby_type_createdonbehalfby_type;
  _createdby_type?: _createdby_type_createdby_type;
  _modifiedby_value?: string;
  _modifiedonbehalfby_value?: string;
  _modifiedonbehalfby_type?: _modifiedonbehalfby_type_modifiedonbehalfby_type;
  _modifiedby_type?: _modifiedby_type_modifiedby_type;
  _ownerid_value?: string;
  _ownerid_type?: _ownerid_type_ownerid_type;
  _owningbusinessunit_value?: string;
  _owningbusinessunit_type?: _owningbusinessunit_type_owningbusinessunit_type;
  _statecode_label?: _statecode_label_statecode_label;
  _statuscode_label?: _statuscode_label_statuscode_label;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
}